# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/properties": {
            "post": {
                "tags": [
                    "Properties"
                ],
                "summary": "Store a new property",
                "operationId": "70bf627c7a3b7593179b22a21956b41e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "location",
                                    "discription",
                                    "number",
                                    "Fridge",
                                    "Bathroom",
                                    "Tank",
                                    "stove",
                                    "Solar",
                                    "parking",
                                    "TotalRooms",
                                    "Roommates",
                                    "Boarding",
                                    "Selling",
                                    "wifi",
                                    "giza",
                                    "Gender",
                                    "price",
                                    "Geo",
                                    "University",
                                    "room",
                                    "Kitchen",
                                    "Toilet",
                                    "out",
                                    "landlord"
                                ],
                                "properties": {
                                    "location": {
                                        "type": "string"
                                    },
                                    "discription": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "Fridge": {
                                        "type": "string"
                                    },
                                    "Bathroom": {
                                        "type": "string"
                                    },
                                    "Tank": {
                                        "type": "string"
                                    },
                                    "stove": {
                                        "type": "string"
                                    },
                                    "Solar": {
                                        "type": "string"
                                    },
                                    "parking": {
                                        "type": "string"
                                    },
                                    "TotalRooms": {
                                        "type": "string"
                                    },
                                    "Roommates": {
                                        "type": "string"
                                    },
                                    "Boarding": {
                                        "type": "string"
                                    },
                                    "Selling": {
                                        "type": "string"
                                    },
                                    "titleland": {
                                        "type": "string"
                                    },
                                    "wifi": {
                                        "type": "string"
                                    },
                                    "giza": {
                                        "type": "string"
                                    },
                                    "Gender": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "Geo": {
                                        "type": "string"
                                    },
                                    "University": {
                                        "type": "string"
                                    },
                                    "room": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Kitchen": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Toilet": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "out": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "landlord": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Property added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property added successfully."
                                        },
                                        "property": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/myproperties": {
            "get": {
                "tags": [
                    "Properties"
                ],
                "summary": "Get all properties created by the authenticated user",
                "operationId": "5f199c4c17505a912da2ca24ed24523e",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved properties",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "name": {
                                                "type": "string"
                                            },
                                            "location": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "like": {
                                                "type": "integer"
                                            },
                                            "Fridge": {
                                                "type": "integer"
                                            },
                                            "WaterTank": {
                                                "type": "stringr"
                                            },
                                            "Solar": {
                                                "type": "string"
                                            },
                                            "Roommates": {
                                                "type": "string"
                                            },
                                            "wifi": {
                                                "type": "string"
                                            },
                                            "geo": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string",
                                                "format": "url",
                                                "example": "http://yourdomain.com/storage/properties/room.jpg"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/properties/{id}": {
            "put": {
                "tags": [
                    "Properties"
                ],
                "summary": "Update an existing property",
                "operationId": "19a6c5948a5a419de858916d8a4ab7c6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location": {
                                        "type": "string"
                                    },
                                    "discription": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "integer"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "room": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Kitchen": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "Toilet": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "out": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Property updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Property updated successfully."
                                        },
                                        "property": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Property not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token to access protected routes",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}